[DONTMERGE] surfna example configuration
This uses a TWRP device tree generated using

 - https://github.com/SebaUbuntu/TWRP-device-tree-generator

And a known working recovery.img as a base.

Though it does not boot.

I guess one would need to investigate with other devices to have a
better understanding of why.

The recovery.img this produces has the same exact configuration and
files, according to unpackbootimg, except for os level, os version and
cmdline. Re-packing with those from the vendor recovery.img changed
nothing.

So, to recap, I have repacked from an unpackimg extraction BOTH this
TWRP and the stock recovery. The repacked stock recovery still works,
any TWRP will not.

The only difference is the ramdisk.